Using proper ivfpq types

+++ cpp/src/distance/neighbors/specializations/detail/ivfpq_search_int8_t_uint64_t.cu
@@ -18,14 +18,14 @@
 
 namespace raft::spatial::knn::ivf_pq::detail {
 
-template void search<float, uint32_t>(const raft::device_resources&,
-                                      const search_params&,
-                                      const index<uint32_t>&,
-                                      const float*,
-                                      uint32_t,
-                                      uint32_t,
-                                      uint32_t*,
-                                      float*,
-                                      rmm::mr::device_memory_resource*);
+template void search<int8_t, uint64_t>(const raft::device_resources&,
+                                       const search_params&,
+                                       const index<uint64_t>&,
+                                       const int8_t*,
+                                       uint32_t,
+                                       uint32_t,
+                                       uint64_t*,
+                                       float*,
+                                       rmm::mr::device_memory_resource*);
 
 }  // namespace raft::spatial::knn::ivf_pq::detail
